requirements.txt
langchain
langchain-google-genai
discord.py
python-dotenv